Move behaviours to its own package
While these are technically listeners, they are a bit special as they
are linked to variables that have to be caught by the primary listener
system.
